Project Structure refactoring (components moves and renames)
