+++ src/components/Chat/ChatContent/ChatHeader.tsx
@@ -2,11 +2,11 @@ import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { shallow } from 'zustand/shallow';
 import useStore from '@store/store';
-import ConfigMenu from '@components/ConfigMenu';
+import ConfigMenu from '@components/Chat/ChatConfigMenu';
 import { ChatInterface, ConfigInterface } from '@type/chat';
 import { _defaultChatConfig, supportedModels } from '@constants/chat';
 
-const ChatTitle = React.memo(() => {
+const ChatHeader = React.memo(() => {
   const { t } = useTranslation('model');
   const config = useStore(
     (state) =>
@@ -89,4 +89,4 @@ const ChatTitle = React.memo(() => {
   );
 });
 
-export default ChatTitle;
+export default ChatHeader;
