+++ src/components/Chat/SearchBar/index.ts
