+++ src/components/Chat/ShareGPT/index.ts
