+++ src/components/Chat/StopGeneratingButton/StopGeneratingButton.tsx
