--- src/components/ConfigMenu/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ConfigMenu';
