+++ src/components/Menu/SettingsMenu/DefaultChatConfigMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DefaultChatConfigMenu';
