+++ src/components/Menu/SettingsMenu/EnterToSubmitToggle.tsx
@@ -1,28 +1,28 @@
 import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import useStore from '@store/store';
-import Toggle from '@components/Toggle';
+import Toggle from '@components/Menu/Toggle';
 
-const AdvancedModeToggle = () => {
+const EnterToSubmitToggle = () => {
   const { t } = useTranslation();
 
-  const setAdvancedMode = useStore((state) => state.setAdvancedMode);
+  const setEnterToSubmit = useStore((state) => state.setEnterToSubmit);
 
   const [isChecked, setIsChecked] = useState<boolean>(
-    useStore.getState().advancedMode
+    useStore.getState().enterToSubmit
   );
 
   useEffect(() => {
-    setAdvancedMode(isChecked);
+    setEnterToSubmit(isChecked);
   }, [isChecked]);
 
   return (
     <Toggle
-      label={t('advancedMode') as string}
+      label={t('enterToSubmit') as string}
       isChecked={isChecked}
       setIsChecked={setIsChecked}
     />
   );
 };
 
-export default AdvancedModeToggle;
+export default EnterToSubmitToggle;
