+++ src/components/Menu/SettingsMenu/SettingsMenu.tsx
@@ -5,20 +5,20 @@ import useCloudAuthStore from '@store/cloud-auth-store';
 
 import PopupModal from '@components/PopupModal';
 import SettingIcon from '@icon/SettingIcon';
-import ThemeSwitcher from '@components/SettingsMenu/ThemeSwitcher';
-import LanguageSelector from '@components/LanguageSelector';
+import ThemeSwitcher from '@components/Menu/SettingsMenu/ThemeSwitcher';
+import LanguageSelector from '@components/Menu/SettingsMenu/LanguageSelector';
 import AutoTitleToggle from './AutoTitleToggle';
 import AdvancedModeToggle from './AdvencedModeToggle';
 import InlineLatexToggle from './InlineLatexToggle';
 import EnterToSubmitToggle from './EnterToSubmitToggle';
 import RequestTokensCountToggle from './RequestTokensCountToggle'
-import { TotalTokenCostToggle } from '../Menu/MenuOptions/TotalTokenCost';
+import { TotalTokenCostToggle } from '../MenuOptions/TotalTokenCost';
 import ChatNamesAsPageTitlesToggle from './ChatNamesAsPageTitlesToggle';
-import PromptLibraryMenu from '@components/PromptLibraryMenu';
-import ChatConfigMenu from '@components/ChatConfigMenu';
+import PromptLibraryMenu from '@components/Menu/SettingsMenu/PromptLibraryMenu';
+import ChatConfigMenu from '@components/Menu/SettingsMenu/DefaultChatConfigMenu';
 import ApiButton from './ApiButton';
 import ClearConversation from '@components/Menu/MenuOptions/ClearConversation';
-import ImportExportChat from '@components/ImportExportChat';
+import ImportExportChat from '@components/Chat/ImportExportChat';
 import { _defaultChatConfig, _defaultSystemMessage } from '@constants/chat';
